A prototype youtube module.
This is really just a proof of concept for using socket.io.
The server side youtube module code is all written in app.js
and the client side stuff needs to be cleaned up. But, it's
sending a video id from the tail will play the video on
the head.